[chore] : Networking Code 추가 및 수정
