--- MarketApp/MarketApp/Network/AffinityAPI.swift
@@ -1,15 +1,0 @@
-//
-//  Affinity.swift
-//  MarketApp
-//
-//  Created by 서원지 on 2023/02/13.
-//
-
-import Foundation
-
-enum AffinityAPI {
-    
-    static let MainShoes = String("/CreamMainShoes")
-    
-}
-
